👩‍❤️‍💋‍👩🐸 Checkpoint ./commands/close.js:850652/207 ./server.js:850652/8675

+++ commands/close.js
@@ -7,7 +7,7 @@ exports.run = (client, message, args, tools) => {
   const db = low(stepsdb);
   if (!message.channel.name.includes("ticket-")) return;
   var data = db.find({ id: message.author.id }).value();
-  if (!data) return;
-  db.remove({ id: message.author.id }).write();
+  if ( message.member.hasPermission("ADMINISTRATOR") == false || !data ) return;
+  if(data) db.remove({ id: message.author.id }).write();
   message.channel.delete();
 };
